refactor(gitconfig): fix namespace
