+++ src/gitconfig/mod.rs
@@ -1,2 +1,4 @@
-mod value;
-pub use self::value::{Value, Map, Entry};
+pub mod value;
+pub mod map;
+pub use self::value::Value;
+pub use self::map::Map;
